handleError * refactorings
