+++ frontend/src/app/trees/_services/tree.service.spec.ts
@@ -3,15 +3,16 @@ import { TreesService } from './trees.service';
 import { Http, HttpModule } from '@angular/http';
 import { RouterTestingModule } from '@angular/router/testing';
 import * as sinon from 'sinon';
+import { UtilService } from '../../_services/util.service';
 
 describe('TreesService', () => {
   let service: TreesService;
 
   beforeEach(() => {
-    service = new TreesService(null, null);
+    service = new TreesService(null, null, null);
 
     TestBed.configureTestingModule({
-      providers: [{ provide: TreesService }]
+      providers: [UtilService, { provide: TreesService }]
     });
   });
 });
